srv_mtga_be/pyproject.toml
[tool.poetry]
name = "mtg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.10.8"
grpcio = "^1.51.0"
grpcio-tools = "^1.51.0"
pytest-cov = "^4.0.0"
pymongo = "^4.3.3"
load-dotenv = "^0.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
